Implement SceneSwitchController as IChangable
SceneSwitchController takes care of switching bewteen parts of the scene. Implementation relies on Fungus scripting
and actions therefore all should be saved and loaded just fine.

There are many hacks in Fungus. This is because we do not want to fade-in/out when loading game. That's when we call
different Fungus. If however called from from regular game this node is executed as well.
Things like MoveTo in 0 seconds and FadeIn in 0 seconds do not do anything visually but are necessary when loading the
game.